# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for CruiseBooking",
        "title": "CruiseBooking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/images": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "List images by entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity type",
                        "name": "entity_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Save images for entity",
                "parameters": [
                    {
                        "description": "Image payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SaveImagesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/upload/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload an image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "github_com_cruisebooking_backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码（0 表示成功）",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.ImageItemRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "is_primary": {
                    "description": "是否主图",
                    "type": "boolean"
                },
                "sort_order": {
                    "description": "排序",
                    "type": "integer"
                },
                "url": {
                    "description": "图片地址",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.SaveImagesRequest": {
            "type": "object",
            "required": [
                "entity_id",
                "entity_type"
            ],
            "properties": {
                "entity_id": {
                    "description": "实体 ID",
                    "type": "integer"
                },
                "entity_type": {
                    "description": "实体类型",
                    "type": "string"
                },
                "images": {
                    "description": "图片列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler.ImageItemRequest"
                    }
                }
            }
        }
    }
}